Migration of https://github.com/airbytehq/airbyte/pull/23002
