+++ airbyte-webapp/src/packages/cloud/cloudRoutes.tsx
@@ -17,22 +17,27 @@ import { CompleteOauthRequest } from "views/CompleteOauthRequest";
 import { CloudRoutes } from "./cloudRoutePaths";
 import { LDExperimentServiceProvider } from "./services/thirdParty/launchdarkly";
 import { VerifyEmailAction } from "./views/FirebaseActionRoute";
-import { RoutePaths, DestinationPaths } from "../../pages/routePaths";
+import { RoutePaths, DestinationPaths, SourcePaths } from "../../pages/routePaths";
 
 const CloudMainView = React.lazy(() => import("packages/cloud/views/layout/CloudMainView"));
 const WorkspacesPage = React.lazy(() => import("packages/cloud/views/workspaces"));
 const Auth = React.lazy(() => import("packages/cloud/views/auth"));
 const CreditsPage = React.lazy(() => import("packages/cloud/views/credits"));
+const SpeakeasyRedirectPage = React.lazy(() => import("pages/SpeakeasyRedirectPage"));
 
 const ConnectionsRoutes = React.lazy(() => import("pages/connections/ConnectionsRoutes"));
-const CreateConnectionPage = React.lazy(() => import("pages/connections/CreateConnectionPage"));
+
 const AllDestinationsPage = React.lazy(() => import("pages/destination/AllDestinationsPage"));
 const CreateDestinationPage = React.lazy(() => import("pages/destination/CreateDestinationPage"));
 const DestinationItemPage = React.lazy(() => import("pages/destination/DestinationItemPage"));
 const DestinationOverviewPage = React.lazy(() => import("pages/destination/DestinationOverviewPage"));
 const DestinationSettingsPage = React.lazy(() => import("pages/destination/DestinationSettingsPage"));
-const SourcesPage = React.lazy(() => import("pages/SourcesPage"));
-const SpeakeasyRedirectPage = React.lazy(() => import("pages/SpeakeasyRedirectPage"));
+
+const AllSourcesPage = React.lazy(() => import("pages/source/AllSourcesPage"));
+const CreateSourcePage = React.lazy(() => import("pages/source/CreateSourcePage"));
+const SourceItemPage = React.lazy(() => import("pages/source/SourceItemPage"));
+const SourceOverviewPage = React.lazy(() => import("pages/source/SourceOverviewPage"));
+const SourceSettingsPage = React.lazy(() => import("pages/source/SourceSettingsPage"));
 
 const CloudSettingsPage = React.lazy(() => import("./views/settings/CloudSettingsPage"));
 const DefaultView = React.lazy(() => import("./views/DefaultView"));
@@ -55,13 +60,19 @@ const MainRoutes: React.FC = () => {
         <Route path={RoutePaths.Destination}>
           <Route index element={<AllDestinationsPage />} />
           <Route path={DestinationPaths.NewDestination} element={<CreateDestinationPage />} />
-          <Route path={DestinationPaths.NewConnection} element={<CreateConnectionPage />} />
           <Route path={DestinationPaths.Root} element={<DestinationItemPage />}>
-            <Route path={DestinationPaths.Settings} element={<DestinationSettingsPage />} />
             <Route index element={<DestinationOverviewPage />} />
+            <Route path={DestinationPaths.Settings} element={<DestinationSettingsPage />} />
           </Route>
         </Route>
-        <Route path={`${RoutePaths.Source}/*`} element={<SourcesPage />} />
+        <Route path={RoutePaths.Source}>
+          <Route index element={<AllSourcesPage />} />
+          <Route path={SourcePaths.NewSource} element={<CreateSourcePage />} />
+          <Route path={SourcePaths.Root} element={<SourceItemPage />}>
+            <Route index element={<SourceOverviewPage />} />
+            <Route path={SourcePaths.Settings} element={<SourceSettingsPage />} />
+          </Route>
+        </Route>
         <Route path={`${RoutePaths.Connections}/*`} element={<ConnectionsRoutes />} />
         <Route path={`${RoutePaths.Settings}/*`} element={<CloudSettingsPage />} />
         <Route path={CloudRoutes.Credits} element={<CreditsPage />} />
