--- airbyte-webapp/src/pages/SourcesPage/index.tsx
@@ -1,3 +1,0 @@
-import { SourcesPage } from "./SourcesPage";
-
-export default SourcesPage;
