--- airbyte-webapp/src/pages/SourcesPage/pages/SourceItemPage/index.tsx
@@ -1,3 +1,0 @@
-import SourceItemPage from "./SourceItemPage";
-
-export default SourceItemPage;
