+++ airbyte-webapp/src/pages/routePaths.tsx
@@ -23,5 +23,10 @@ export enum DestinationPaths {
   Root = ":id/*", // currently our tabs rely on this * wildcard to detect which tab is currently active
   Settings = "settings",
   NewDestination = "new-destination",
-  NewConnection = "new-connection",
+}
+
+export enum SourcePaths {
+  Root = ":id/*", // currently our tabs rely on this * wildcard to detect which tab is currently active
+  Settings = "settings",
+  NewSource = "new-source",
 }
