+++ airbyte-webapp/src/pages/source/CreateSourcePage/index.tsx
@@ -1,0 +1,1 @@
+export { CreateSourcePage as default } from "./CreateSourcePage";
