+++ airbyte-webapp/src/pages/source/SourceItemPage/index.tsx
@@ -1,0 +1,1 @@
+export { SourceItemPage as default } from "./SourceItemPage";
