+++ airbyte-webapp/src/pages/source/SourceOverviewPage/SourceConnectionTable.module.scss
@@ -1,4 +1,4 @@
-@use "../../../../../scss/variables";
+@use "scss/variables";
 
 .content {
   padding: 0 variables.$spacing-xl;
