+++ airbyte-webapp/src/pages/source/SourceOverviewPage/index.tsx
@@ -1,0 +1,1 @@
+export { SourceOverviewPage as default } from "./SourceOverviewPage";
