+++ airbyte-webapp/src/pages/source/SourceSettingsPage/SourceSettingsPage.module.scss
@@ -1,4 +1,4 @@
-@use "../../../../../scss/variables";
+@use "scss/variables";
 
 .content {
   width: 80%;
