+++ airbyte-webapp/src/pages/source/SourceSettingsPage/SourceSettingsPage.tsx
@@ -2,29 +2,25 @@ import React, { useCallback, useEffect, useMemo } from "react";
 import { FormattedMessage } from "react-intl";
 
 import { ConnectionConfiguration } from "core/domain/connection";
-import { SourceRead, WebBackendConnectionListItem } from "core/request/AirbyteClient";
 import { useTrackPage, PageTrackingCodes } from "hooks/services/Analytics";
 import { useFormChangeTrackerService, useUniqueFormId } from "hooks/services/FormChangeTracker";
 import { useDeleteSource, useInvalidateSource, useUpdateSource } from "hooks/services/useSourceHook";
 import { useDeleteModal } from "hooks/useDeleteModal";
-import { useSourceDefinition } from "services/connector/SourceDefinitionService";
 import { useGetSourceDefinitionSpecification } from "services/connector/SourceDefinitionSpecificationService";
 import { ConnectorCard } from "views/Connector/ConnectorCard";
 import { useDocumentationPanelContext } from "views/Connector/ConnectorDocumentationLayout/DocumentationPanelContext";
 
-import styles from "./SourceSettings.module.scss";
+import styles from "./SourceSettingsPage.module.scss";
+import { useSourceOverviewContext } from "../SourceOverviewPage/sourceOverviewContext";
 
-interface SourceSettingsProps {
-  currentSource: SourceRead;
-  connectionsWithSource: WebBackendConnectionListItem[];
-}
-
-const SourceSettings: React.FC<SourceSettingsProps> = ({ currentSource, connectionsWithSource }) => {
+export const SourceSettingsPage: React.FC = () => {
   const { mutateAsync: updateSource } = useUpdateSource();
   const { mutateAsync: deleteSource } = useDeleteSource();
   const { setDocumentationPanelOpen } = useDocumentationPanelContext();
   const formId = useUniqueFormId();
   const { clearFormChange } = useFormChangeTrackerService();
+
+  const { source, sourceDefinition, connections } = useSourceOverviewContext();
 
   useTrackPage(PageTrackingCodes.SOURCE_ITEM_SETTINGS);
   useEffect(() => {
@@ -33,10 +29,9 @@ const SourceSettings: React.FC<SourceSettingsProps> = ({ currentSource, connecti
     };
   }, [setDocumentationPanelOpen]);
 
-  const sourceDefinitionSpecification = useGetSourceDefinitionSpecification(currentSource.sourceDefinitionId);
+  const sourceDefinitionSpecification = useGetSourceDefinitionSpecification(source.sourceDefinitionId);
 
-  const sourceDefinition = useSourceDefinition(currentSource.sourceDefinitionId);
-  const reloadSource = useInvalidateSource(currentSource.sourceId);
+  const reloadSource = useInvalidateSource(source.sourceId);
 
   const onSubmit = async (values: {
     name: string;
@@ -45,30 +40,30 @@ const SourceSettings: React.FC<SourceSettingsProps> = ({ currentSource, connecti
   }) => {
     await updateSource({
       values,
-      sourceId: currentSource.sourceId,
+      sourceId: source.sourceId,
     });
   };
 
   const onDelete = useCallback(async () => {
     clearFormChange(formId);
-    await deleteSource({ connectionsWithSource, source: currentSource });
-  }, [clearFormChange, connectionsWithSource, currentSource, deleteSource, formId]);
+    await deleteSource({ connectionsWithSource: connections, source });
+  }, [clearFormChange, connections, source, deleteSource, formId]);
 
   const modalAdditionalContent = useMemo<React.ReactNode>(() => {
-    if (connectionsWithSource.length === 0) {
+    if (connections.length === 0) {
       return null;
     }
     return (
       <p>
-        <FormattedMessage id="tables.affectedConnectionsOnDeletion" values={{ count: connectionsWithSource.length }} />
-        {connectionsWithSource.map((connection) => (
+        <FormattedMessage id="tables.affectedConnectionsOnDeletion" values={{ count: connections.length }} />
+        {connections.map((connection) => (
           <React.Fragment key={connection.connectionId}>
             - <strong>{`${connection.name}\n`}</strong>
           </React.Fragment>
         ))}
       </p>
     );
-  }, [connectionsWithSource]);
+  }, [connections]);
 
   const onDeleteClick = useDeleteModal("source", onDelete, modalAdditionalContent);
 
@@ -82,7 +77,7 @@ const SourceSettings: React.FC<SourceSettingsProps> = ({ currentSource, connecti
         availableConnectorDefinitions={[sourceDefinition]}
         selectedConnectorDefinitionSpecification={sourceDefinitionSpecification}
         selectedConnectorDefinitionId={sourceDefinitionSpecification.sourceDefinitionId}
-        connector={currentSource}
+        connector={source}
         reloadConfig={reloadSource}
         onSubmit={onSubmit}
         onDeleteClick={onDeleteClick}
@@ -90,5 +85,3 @@ const SourceSettings: React.FC<SourceSettingsProps> = ({ currentSource, connecti
     </div>
   );
 };
-
-export default SourceSettings;
