+++ airbyte-webapp/src/pages/source/SourceSettingsPage/index.tsx
@@ -1,0 +1,1 @@
+export { SourceSettingsPage as default } from "./SourceSettingsPage";
